remove AbstractGarbageCollectionStrategy
